Make prefetcher reads async
Right now prefetcher reads are a blocking API. That's a bit weird since
they're called from an async interface (the file system). It's not a
problem right now because the file system async interface in turn is
only called from `block_on` threads in fuser. But if we want to change
to a real async runtime that multiplexes futures across threads, we need
to remove blocking APIs like this one.

Signed-off-by: James Bornholt <bornholt@amazon.com>