Add a proc macro to log `fs` failures automatically
This is a followup to #405 that automatically emits log messages when
annotated functions in `fs` fail (return Err). The macro allows two
different verbosities of log messages depending on whether an error is
"expected" or not, which is defined by a new `ExpectedError` trait. It's
intended to be used in a context where a tracing span already exists to
capture important fields/arguments if necessary, so doesn't allow
customizing the log message.

Signed-off-by: James Bornholt <bornholt@amazon.com>