Add system/process resource usage metrics
This helps us understand system-level CPU and memory pressure and
process-level CPU and memory usage. We bolt this new metrics monitoring
into our existing 5-second metrics cadence, but emit them only when we
emit at least one other metric, to avoid spamming logs indefinitely with
these messages when metric logging is enabled.

Failures of metrics logging should not prevent Mountpoint from running
or staying alive, so we're careful to absorb failures on this path and
keep trucking.

For this initial version I chose to make this on-by-default with no way
to turn it off. We could consider exposing a flag for this if we think
it's useful.

Signed-off-by: James Bornholt <bornholt@amazon.com>